// Banking.NgWeb/src/app/model/Bank/customer.ts
export interface ICustomer {
  FirstName: string;
  LastName: string;
  FullName: string;
  ContactNumber: number;
  Mail: string;
  Dob: Date;
  Gender: string;
  Address: string;
  EmployeeId: number;
  CustomerId: number;
  Number: number;
  Password: any;
  Balance: number;
  AccountType: string;
  AccountTypeId: string;
  Status: boolean;
  OpenDate: string;
  Role: string;

}